assignment-2/clarity/pyproject.toml
[tool.poetry]
name = "python_server"
version = "0.1.0"
description = ""
authors = ["Harry Newton <hazzaprime@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
newspaper3k = "^0.2.8"
aiofiles = "^0.5.0"
sanic = "^20.9.0"
nltk = "^3.5"
spacy = "^2.3.2"
pandas = "^1.1.3"
sklearn = "^0.0"
dateparser = "^0.7.6"
newsapi-python = "^0.2.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
